+++ CHANGELOG.md
@@ -157,6 +157,8 @@
 * Added OnFire and OnReload to item scripts (for weapons only)
 * FacingSprite and BillboardSprite now has a "Bright" flag that draws at full brightness
 * Fixed heading font not applying to modal headings
+* Fixed BlankSceneController not added to script execution order
+* Implemented Generic Translate Script
 * Weapon ViewModel can be forced to always or never wait for lock time to do effect [Experimental]
 * Added subtitle convenience methods from Shattered 2 [Experimental]
 * Extended projectile and explosion handling for weapons [WIP]